Create standalone landfire test module
