+++ tests/test_calibration_module.py
@@ -9,13 +9,11 @@ import numpy as np
 from pathlib import Path
 import math
 import matplotlib.pyplot as plt
-import geojson
 
 from duet_tools.calibration import (
     DuetRun,
     Targets,
     FuelParameter,
-    LandfireQuery,
     import_duet,
     assign_targets,
     set_fuel_parameter,
@@ -23,12 +21,10 @@ from duet_tools.calibration import (
     set_moisture,
     set_height,
     calibrate,
-    assign_targets_from_sb40,
     _maxmin_calibration,
 )
 
 from duet_tools.utils import read_dat_to_array
-from duet_tools.landfire import query_landfire
 
 TEST_DIR = Path(__file__).parent
 TMP_DIR = TEST_DIR / "tmp"
@@ -357,65 +353,6 @@ class TestCalibrate:
         assert calibrated_duet.density.shape[0] == 2
         assert np.max(np.sum(calibrated_duet.density, axis=0)) == np.float32(2.0)
         assert np.min(np.sum(calibrated_duet.density, axis=0)) == np.float32(0.5)
-
-
-class TestLandfireTargets:
-    @classmethod
-    def get_geojson(self):
-        with open(DATA_DIR / "anderson_butte.geojson") as fid:
-            sample_geojson = geojson.load(fid)
-        return geojson.Polygon(sample_geojson["features"][0]["geometry"]["coordinates"])
-
-    def test_query_landfire(self):
-        sample_aoi = self.get_geojson()
-        query = query_landfire(
-            area_of_interest=sample_aoi, directory=TMP_DIR, input_epsg=4326
-        )
-        assert isinstance(query, LandfireQuery)
-        assert isinstance(query.fuel_types, np.ndarray)
-        assert isinstance(query.density, np.ndarray)
-        assert isinstance(query.moisture, np.ndarray)
-        assert isinstance(query.height, np.ndarray)
-
-    def test_assign_targets_from_sb40(self):
-        sample_aoi = self.get_geojson()
-        query = query_landfire(
-            area_of_interest=sample_aoi, directory=TMP_DIR, input_epsg=4326
-        )
-        # test just grass density
-        grass_density = assign_targets_from_sb40(query, "grass", "density")
-        assert isinstance(grass_density, Targets)
-        assert grass_density.method == "maxmin"
-        assert grass_density.args == ["max", "min"]
-        assert len(grass_density.targets) == 2
-        # test fuel and parameter with only one value
-        with pytest.warns(UserWarning):
-            grass_moisture = assign_targets_from_sb40(query, "grass", "moisture")
-        grass_moisture = assign_targets_from_sb40(
-            query, "grass", "moisture", method="constant"
-        )
-        # get the rest of the fuels and params with maxmin
-        grass_height = assign_targets_from_sb40(query, "grass", "height")
-        litter_density = assign_targets_from_sb40(query, "litter", "density")
-        litter_moisture = assign_targets_from_sb40(query, "litter", "density")
-        litter_height = assign_targets_from_sb40(query, "litter", "density")
-        all_density = assign_targets_from_sb40(query, "all", "density")
-        all_moisture = assign_targets_from_sb40(query, "all", "moisture")
-        all_height = assign_targets_from_sb40(query, "all", "height")
-        # test a couple with meansd
-        grass_density_meansd = assign_targets_from_sb40(
-            query, "grass", "density", method="meansd"
-        )
-        all_height = assign_targets_from_sb40(query, "all", "height", "meansd")
-        # test wrong inputs
-        with pytest.raises(ValueError):
-            grass_height = assign_targets_from_sb40(query, "both", "height")
-        with pytest.raises(ValueError):
-            grass_height = assign_targets_from_sb40(query, "all", "moist")
-        with pytest.raises(ValueError):
-            grass_height = assign_targets_from_sb40(
-                query, "all", "height", method="minmax"
-            )
 
 
 def plot_array(x, title):
